+++ src/integrations/lottie/mod.rs
@@ -1,5 +1,10 @@
 mod asset_loader;
 
+pub(crate) mod render;
+
+pub mod asset;
+pub use asset::{VelloLottie, VelloLottieHandle};
+
 mod systems;
 #[cfg(feature = "experimental-dotLottie")]
 pub(crate) use systems::spawn_playheads;
@@ -23,3 +28,77 @@ pub use playhead::Playhead;
 
 mod theme;
 pub use theme::Theme;
+
+use crate::{debug::DebugVisualizations, CoordinateSpace};
+use bevy::prelude::*;
+
+#[cfg(feature = "lottie")]
+#[derive(Bundle, Default)]
+pub struct VelloLottieBundle {
+    /// Asset data to render
+    pub asset: VelloLottieHandle,
+    /// How the asset is positioned relative to its [`Transform`].
+    pub asset_anchor: VelloLottieAnchor,
+    /// The coordinate space in which this vector should be rendered.
+    pub coordinate_space: CoordinateSpace,
+    /// A transform to apply to this vector
+    pub transform: Transform,
+    /// Whether to render debug visualizations
+    pub debug_visualizations: DebugVisualizations,
+    /// User indication of whether an entity is visible. Propagates down the entity hierarchy.
+    pub visibility: Visibility,
+}
+
+/// Describes how the asset is positioned relative to its [`Transform`]. It defaults to [`VelloAssetAnchor::Center`].
+#[derive(Component, Default, Clone, Copy, PartialEq, Eq)]
+pub enum VelloLottieAnchor {
+    /// Bounds start from the render position and advance up and to the right.
+    BottomLeft,
+    /// Bounds start from the render position and advance up.
+    Bottom,
+    /// Bounds start from the render position and advance up and to the left.
+    BottomRight,
+
+    /// Bounds start from the render position and advance right.
+    Left,
+    /// Bounds start from the render position and advance equally on both axes.
+    #[default]
+    Center,
+    /// Bounds start from the render position and advance left.
+    Right,
+
+    /// Bounds start from the render position and advance down and to the right.
+    TopLeft,
+    /// Bounds start from the render position and advance down.
+    Top,
+    /// Bounds start from the render position and advance down and to the left.
+    TopRight,
+}
+
+impl VelloLottieAnchor {
+    pub(crate) fn compute(
+        &self,
+        width: f32,
+        height: f32,
+        transform: &GlobalTransform,
+    ) -> GlobalTransform {
+        // Apply positioning
+        let adjustment = match self {
+            Self::TopLeft => Vec3::new(width / 2.0, -height / 2.0, 0.0),
+            Self::Left => Vec3::new(width / 2.0, 0.0, 0.0),
+            Self::BottomLeft => Vec3::new(width / 2.0, height / 2.0, 0.0),
+            Self::Top => Vec3::new(0.0, -height / 2.0, 0.0),
+            Self::Center => Vec3::new(0.0, 0.0, 0.0),
+            Self::Bottom => Vec3::new(0.0, height / 2.0, 0.0),
+            Self::TopRight => Vec3::new(-width / 2.0, -height / 2.0, 0.0),
+            Self::Right => Vec3::new(-width / 2.0, 0.0, 0.0),
+            Self::BottomRight => Vec3::new(-width / 2.0, height / 2.0, 0.0),
+        };
+        let new_translation: Vec3 = (transform.compute_matrix() * adjustment.extend(1.0)).xyz();
+        GlobalTransform::from(
+            transform
+                .compute_transform()
+                .with_translation(new_translation),
+        )
+    }
+}